requirements.txt
#!/usr/bin/env -S pip install -r

lsp-tree-sitter
tree-sitter-make
